+++ functions/run/types.go
@@ -1,8 +1,10 @@
 package run
 
+// LocalProc represents the result of executing a local process
+// from a Starlark script.
 type LocalProc struct {
-	Pid    int
-	Error  string
-	Result string
-	ExitCode int
+	Pid      int64
+	Error    string
+	Result   string
+	ExitCode int64
 }
